+++ aa_forum/tests/test_integration.py
@@ -108,6 +108,37 @@ class TestForumUI(WebTest):
         new_message = Message.objects.last()
         self.assertEqual(new_message.message, "Energy of the Higgs boson")
 
+    def test_should_not_create_new_topic_doe_to_invalid_form(self):
+        """
+        Test should not create new topic due to invalid form
+        :return:
+        """
+
+        # given
+        self.app.set_user(self.user_1001)
+        page = self.app.get(self.board.get_absolute_url())
+
+        # when
+        page = page.click(linkid="aa-forum-btn-new-topic-above-list")
+        form = page.forms["aa-forum-form-new-topic"]
+        # form["subject"] = "Recent Discoveries"  # commented to create an invalid form
+        form["message"] = "Energy of the Higgs boson"
+        page = form.submit()
+
+        # then
+        self.assertEqual(self.board.topics.count(), 0)  # No topic created
+        self.assertTemplateUsed(page, "aa_forum/view/forum/new-topic.html")
+
+        expected_message = (
+            "<h4>Error!</h4>"
+            "<p>Either subject or message is missing. "
+            "Please make sure you enter both fields, "
+            "as both fields are mandatory.</p>"
+        )
+        messages = list(page.context["messages"])
+        self.assertEqual(len(messages), 1)
+        self.assertEqual(str(messages[0]), expected_message)
+
     def test_should_not_create_topic_that_already_exists(self):
         """
         Test should not re-create an existing topic